added file class, added convertToPath function, added convert char array to string function, updated run function
added file class it will be used for file pathing and also file name. added convert to path function which converts path if  \ is written to /path. Added convert char array to string function which will convert char array to string. It could be used for multiple things in this application.  updated run function now user can enter file path from the keyboard and that same path would be converted to the path which do not contains \ only / then that path is printed by message static class.